// src/api/models/update-base-information-req.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CounselPurposeAndNoteDTO } from './counsel-purpose-and-note-dto';

/**
 * 
 * @export
 * @interface UpdateBaseInformationReq
 */
export interface UpdateBaseInformationReq {
    /**
     * 
     * @type {CounselPurposeAndNoteDTO}
     * @memberof UpdateBaseInformationReq
     */
    'counselPurposeAndNote': CounselPurposeAndNoteDTO;
}