// src/api/models/update-independent-life-information-req.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CommunicationDTO } from './communication-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { EvacuationDTO } from './evacuation-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { WalkingDTO } from './walking-dto';

/**
 * 
 * @export
 * @interface UpdateIndependentLifeInformationReq
 */
export interface UpdateIndependentLifeInformationReq {
    /**
     * 
     * @type {CommunicationDTO}
     * @memberof UpdateIndependentLifeInformationReq
     */
    'communication': CommunicationDTO;
    /**
     * 
     * @type {EvacuationDTO}
     * @memberof UpdateIndependentLifeInformationReq
     */
    'evacuation': EvacuationDTO;
    /**
     * 
     * @type {WalkingDTO}
     * @memberof UpdateIndependentLifeInformationReq
     */
    'walking': WalkingDTO;
}