// src/api/models/update-living-information-req.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DrinkingDTO } from './drinking-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { ExerciseDTO } from './exercise-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { MedicationManagementDTO } from './medication-management-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { NutritionDTO } from './nutrition-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { SmokingDTO } from './smoking-dto';

/**
 * 
 * @export
 * @interface UpdateLivingInformationReq
 */
export interface UpdateLivingInformationReq {
    /**
     * 
     * @type {DrinkingDTO}
     * @memberof UpdateLivingInformationReq
     */
    'drinking': DrinkingDTO;
    /**
     * 
     * @type {ExerciseDTO}
     * @memberof UpdateLivingInformationReq
     */
    'exercise': ExerciseDTO;
    /**
     * 
     * @type {MedicationManagementDTO}
     * @memberof UpdateLivingInformationReq
     */
    'medicationManagement': MedicationManagementDTO;
    /**
     * 
     * @type {NutritionDTO}
     * @memberof UpdateLivingInformationReq
     */
    'nutrition': NutritionDTO;
    /**
     * 
     * @type {SmokingDTO}
     * @memberof UpdateLivingInformationReq
     */
    'smoking': SmokingDTO;
}